Implement interpolation between selected tiles for component placement
It's entirely possible that the cursor will move more than 1 tile
between frames, so we need to somehow figure out which tiles were passed over.
I'm creating a linear approximation of the cursor's path and using a
line drawing algorithm to find all of the tiles in between the old and
new positions.